moved policy components to Components to share
